Split update-metrics into three separate functions
