+++ common/loans.ts
@@ -17,39 +17,41 @@ const calculateNewLoan = (investedValue: number, loanTotal: number) => {
   return netValue * LOAN_DAILY_RATE
 }
 
+export const getUserLoanUpdates = (
+  bets: Bet[],
+  contractsById: { [contractId: string]: Contract },
+  portfolio?: PortfolioMetrics | undefined
+) => {
+  if (isUserEligibleForLoan(portfolio)) {
+    const updates = calculateLoanBetUpdates(bets, contractsById).betUpdates
+    return { updates, payout: sumBy(updates, (update) => update.newLoan) }
+  } else {
+    return undefined
+  }
+}
+
 export const getLoanUpdates = (
   users: User[],
   contractsById: { [contractId: string]: Contract },
   portfolioByUser: { [userId: string]: PortfolioMetrics | undefined },
   betsByUser: { [userId: string]: Bet[] }
 ) => {
-  const eligibleUsers = filterDefined(
-    users.map((user) =>
-      isUserEligibleForLoan(portfolioByUser[user.id]) ? user : undefined
-    )
-  )
-
-  const betUpdates = eligibleUsers
-    .map((user) => {
-      const updates = calculateLoanBetUpdates(
+  const userUpdates = filterDefined(
+    users.map((user) => {
+      const result = getUserLoanUpdates(
         betsByUser[user.id] ?? [],
-        contractsById
-      ).betUpdates
-      return updates.map((update) => ({ ...update, user }))
+        contractsById,
+        portfolioByUser[user.id]
+      )
+      return result ? { user, result } : undefined
     })
-    .flat()
-
-  const updatesByUser = groupBy(betUpdates, (update) => update.userId)
-  const userPayouts = Object.values(updatesByUser).map((updates) => {
-    return {
-      user: updates[0].user,
-      payout: sumBy(updates, (update) => update.newLoan),
-    }
-  })
-
+  )
   return {
-    betUpdates,
-    userPayouts,
+    betUpdates: userUpdates.map((u) => u.result.updates).flat(),
+    userPayouts: userUpdates.map(({ user, result: { payout } }) => ({
+      user,
+      payout,
+    })),
   }
 }
