+++ functions/src/scripts/update-metrics.ts
@@ -1,13 +1,18 @@
 import { initAdmin } from './script-init'
 initAdmin()
 
-import { log, logMemory } from '../utils'
-import { updateMetricsCore } from '../update-metrics'
+import { log } from '../utils'
+import { updateContractMetrics } from '../update-contract-metrics'
+import { updateUserMetrics } from '../update-user-metrics'
+import { updateGroupMetrics } from '../update-group-metrics'
 
 async function updateMetrics() {
-  logMemory()
-  log('Updating metrics...')
-  await updateMetricsCore()
+  log('Updating contract metrics...')
+  await updateContractMetrics()
+  log('Updating user metrics...')
+  await updateUserMetrics()
+  log('Updating group metrics...')
+  await updateGroupMetrics()
 }
 
 if (require.main === module) {
