+++ functions/src/utils.ts
@@ -7,6 +7,7 @@ import { Contract } from '../../common/contract'
 import { PrivateUser, User } from '../../common/user'
 import { Group } from '../../common/group'
 import { Post } from '../../common/post'
+import { getFunctionUrl } from '../../common/api'
 
 export const log = (...args: unknown[]) => {
   console.log(`[${new Date().toISOString()}]`, ...args)
@@ -17,6 +18,23 @@ export const logMemory = () => {
   for (const [k, v] of Object.entries(used)) {
     log(`${k} ${Math.round((v / 1024 / 1024) * 100) / 100} MB`)
   }
+}
+
+export const invokeFunction = async (name: string, body?: unknown) => {
+  const response = await fetch(getFunctionUrl(name), {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    method: 'POST',
+    body: JSON.stringify(body ?? {}),
+  })
+
+  const json = await response.json()
+  if (response.ok) return await response.json()
+  else
+    throw new Error(
+      `${response.status} invoking function: ${JSON.stringify(json)}`
+    )
 }
 
 export const revalidateStaticProps = async (
